% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{tiefightR - mouse data}
\format{
A 880 x 10 data frame with continouse response data (numOF_visits_with_Licks).
}
\usage{
data(mouse)
}
\description{
Eleven female C57BL/6J mice were tested. The mice were purchased from Charles River Sulzfeld, Germany and arrived at
the institute at the age of 21 days. All mice were implanted with an RFID chip
(Planet ID, FDX-B transponder according to ISO 11784/85) holding a unique ID for individual differentiation at an age
of 33 or 56 days, a procedure performed under anesthesia (Isofluran). Two hours before the transplantation, all mice were
given an analgesic (Meloxicam). At the start of the first run, the mice were seven months old and weighed between 25.0 and 29.0 g.
At the start of the second run, the mice were 14 month old and weighed between 25.5 and 37.0 g.
}
\examples{
data(mouse)
head(mouse)
}
\keyword{datasets}
