% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhesus.R
\docType{data}
\name{rhesus}
\alias{rhesus}
\title{tiefightR - rhesus data}
\format{
A 240 x 7 data frame with continouse response data (amountDrank).
}
\usage{
data(rhesus)
}
\description{
Six male rhesus macaques (Macaca mulatta) at an age range of 6-19 years (mean = 13.09) living in same-sexual groups
of 2 - 4 at the Cognitive Neuroscience Laboratory, German Primate Centre, took part in this study. Monkeys were housed
in indoor rooms equipped with a multitude of toys and wooden structures as well as natural and artificial light.
The space per monkey exceeded all applicable German and European regulations (Berger et al. 2018). Indoor rooms were
temperature-controlled and connected by a tunnel with rooms at ambient outdoor temperature and lighting, but protected
from precipitation. On test days, monkeys had free access to water for at least 4 hours
(typically much more: for definitions of access to water see (Pfefferle et al. 2018)) and received monkey chow ad libitum.
On non-test days, the monkeys had free access to water and received monkey chow ad libitum,
supplemented with dried fruits, fresh fruits and vegetables. The health of the monkeys was monitored daily by the
animal care staff, veterinarians, and the laboratory researchers who were all highly experienced with these animals.
}
\examples{
data(rhesus)
head(rhesus)
}
\keyword{datasets}
