% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_cicheck.R
\name{tie_cicheck}
\alias{tie_cicheck}
\title{Commodity position and confidence interval check for a discrete number of randomizations}
\usage{
tie_cicheck(
  data = tiefightR::mouse,
  R = NULL,
  ciLvl = 0.95,
  seed = TRUE,
  SV = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  ord = NULL,
  prefLimit = 50,
  compstudy = NULL,
  default = NULL,
  showplot = TRUE,
  showstats = FALSE,
  ylim = c(0.1, 0.35)
)
}
\arguments{
\item{R}{number of maximum randomization steps}

\item{ciLvl}{Level of confidence (default: 0.95)}

\item{seed}{TRUE/FALSE for constant seeding}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{ord}{item category order}

\item{prefLimit}{preference limit for binarization threshold}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{showplot}{show the errorplot with confidence intervals}

\item{showstats}{calculate ANOVA1 and Tukey's test for the commodities}

\item{dat}{imported raw data (should be binary, if not, will be binarized automatically)}
}
\value{
Exports random binarize response for distance cutoff selection
}
\description{
The \code{tie_cicheck} is a wrapper for checking the confidence intervals for data with ties. The function calculates the worth
values for a specific number of randomizations and reports the confidence intervals for the commodity means.
}
