% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_cores.R
\name{tie_cores}
\alias{tie_cores}
\title{CPU Core Detection}
\usage{
tie_cores()
}
\value{
No of CPUs on your machine
}
\description{
The \code{tie_cores} detects the number of available CPUs for parallel computing. Don't overdo it!
}
