% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_cutoff.R
\name{tie_cutoff}
\alias{tie_cutoff}
\title{Cutoff determination for increasing number of randomizations}
\usage{
tie_cutoff(
  data = tiefightR::mouse,
  R = 50,
  ciLvl = 0.95,
  cutoff = 0.1,
  cpus = 2,
  SV = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  ord = NULL,
  prefLimit = 50,
  compstudy = NULL,
  default = NULL,
  showplot = FALSE,
  showCutoff = FALSE
)
}
\arguments{
\item{R}{number of maximum randomization steps}

\item{cutoff}{Percent cutoff level (default: 0.10) - means CI range < than cutoff value}

\item{cpus}{No. of used local CPUs for parallel computing (you should have more than 2)}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{ord}{item category order}

\item{prefLimit}{preference limit for binarization threshold}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{showplot}{show the plot for randomization cutoff determination}

\item{showCutoff}{show vertical line of the cutoff}

\item{dat}{imported raw data (should be binary, if not, will be binarized automatically)}

\item{standardize}{standardize on the maximum CI value?}
}
\value{
Exports cutoff value and plots
}
\description{
The \code{tie_cutoff} function calculates the mean Euclidean distance between commodity worth values. This becomes
relevant when ties are present in the data. Depending on how ties are resolved (see prefLimit argument in the function),
the items' position will change a lot. Since their relative positions are a function of the number of ties, more
randomizations will stabilize their means and thus commodity positions. Increasing the number of randomizations usually leads
not only to a stabilized mean but also to smaller confidence intervals. By defining a relative cutoff (e.g., 5 or 10\%) for the
range of the CIs regarding the maximum range in the data, a cutoff for the number of randomizations can be found.
}
