% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_import.R
\name{tie_import}
\alias{tie_import}
\title{Import Function}
\usage{
tie_import(path = NULL, valenceset = NULL)
}
\arguments{
\item{path}{path to the raw data}

\item{valenceset}{subset filtering argument}
}
\value{
data.frame with the filtered subset
}
\description{
The \code{tie_import} function loads the raw data
}
