% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_intrans.R
\name{tie_intrans}
\alias{tie_intrans}
\title{Calculate Intransitivity}
\usage{
tie_intrans(mydata = NULL)
}
\arguments{
\item{mydata}{input data frame}
}
\value{
intranscount intransitivity counts
}
\description{
The \code{tie_intrans} function loads the raw data
}
