% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_rwalk.R
\name{tie_rwalk}
\alias{tie_rwalk}
\title{Tie random walk function}
\usage{
tie_rwalk(
  dat = NULL,
  SV = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  ord = NULL,
  prefLimit = 50,
  setseed = FALSE,
  compstudy = NULL,
  default = NULL,
  R = NULL
)
}
\arguments{
\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{ord}{item category order}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{xdata}{imported (binarized) data frame}
}
\value{
Exports random binarize response for distance cutoff selection
}
\description{
The \code{tie_rwalk} function prepares binary and continuous data form import into the tiefightR analysis.
}
