% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_rwrap.R
\name{tie_rwrap}
\alias{tie_rwrap}
\title{Tie random wrapper for the rwalk function}
\usage{
tie_rwrap(
  dat = NULL,
  SV = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  ord = NULL,
  default = NULL,
  prefLimit = 50,
  setseed = FALSE,
  compstudy = NULL,
  R = 2
)
}
\arguments{
\item{dat}{imported raw data (should be binary, if not, will be binarized automatically)}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{ord}{item category order}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{prefLimit}{preference limit for binarization threshold}

\item{setseed}{TRUE/FALSE for seeding}

\item{R}{number of maximum randomization steps}

\item{showplotNo}{shows plots: 1 = mean Euclidean distance over R and 2 = same, but SD}
}
\value{
Exports random binarize response for distance cutoff selection
}
\description{
The \code{tie_rwrap} is a wrapper function for tie_rwalk and uses a number of R randomization steps to calculate the mean
Euclidean distance between worth values. The resulting curve will tell something about tie stability in the data and can
be used for the definition of a randomization step cut-off value.
}
