% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_sim.R
\name{tie_sim}
\alias{tie_sim}
\title{Tie Simulation Function}
\usage{
tie_sim(
  xdata = NULL,
  R = 2,
  SV = "side_img1",
  RF = "img1",
  CF = "img2",
  id = "ID",
  RV = "pref_img1",
  intrans = TRUE,
  compstudy = "LagreValenceRange_SpringSchool",
  default = "War",
  cpus = 2,
  ord = NULL,
  v1 = NULL
)
}
\arguments{
\item{xdata}{imported (binarized) data frame}

\item{R}{No. of randomization steps}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{intrans}{calculate intransitivities (calculation intense!)}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{cpus}{No. of used local CPUs for parallel computing (you should have more than 2)}

\item{ord}{item category order}

\item{v1}{testing variable (can be one item from the item list)}
}
\value{
data.frame with the simulation results
}
\description{
The \code{tie_sim} function starts a simulation of item pairings and introduces random pairs for the remaining
combinations. In parallel, the intransitivity of triple pairings can be calculated to estimate their position quality.
Good transitiviy and massed localization in a position will improve the Likelihood of a good fit for the item.
}
