% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_simrep.R
\name{tie_simrep}
\alias{tie_simrep}
\title{Simulation report}
\usage{
tie_simrep(res = NULL, v1 = NULL, path = NULL)
}
\arguments{
\item{res}{result or output from the simulation}

\item{v1}{test variable}

\item{path}{path to where the report shall be stored (inluding plots)}
}
\value{
A frequency table for item positions during the simulation; Tukey's HSD Test for positions; Position Bubble Plot
}
\description{
The \code{tie_simrep} function prepares a frequency table and plots for analysing the simulation output. Can also be used
for saving the output to file when a path is provided.
}
