% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_test.R
\name{tie_test}
\alias{tie_test}
\title{Test Function}
\usage{
tie_test(
  xdata = NULL,
  R = NULL,
  intrans = TRUE,
  compstudy = NULL,
  default = NULL,
  ord = NULL,
  seed = TRUE,
  testme = NULL,
  against = NULL
)
}
\arguments{
\item{xdata}{imported (binarized) data frame}

\item{R}{number of randomizations}

\item{intrans}{calculate intransitivities (calculation intense!)}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{ord}{item category order}

\item{seed}{BOOLEAN; set a random seed TRUE/FALSE?}

\item{testme}{test this variable against...}

\item{against}{any other single variable from the commodity list (or combinations); vectorize if necessary c("item1","item2")}
}
\value{
A frequency table for item positions during the simulation; Tukey's HSD Test for positions; Position Bubble Plot
}
\description{
The \code{tie_test} function can be used for individual item testing.
}
