% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_worth.R
\name{tie_worth}
\alias{tie_worth}
\title{Main preference function}
\usage{
tie_worth(
  xdata = NULL,
  esti = "worth",
  SV = "side_img1",
  RF = "img1",
  CF = "img2",
  id = "ID",
  RV = "pref_img1",
  default = NULL,
  showplot = FALSE,
  intrans = FALSE,
  compstudy = NULL,
  ordn = NULL,
  r1 = NULL,
  r2 = NULL,
  ymin = 0,
  ymax = 0.5
)
}
\arguments{
\item{xdata}{imported (binarized) data frame}

\item{esti}{worth estimator (default, "worth", alt: "estimator")}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{showplot}{show worth plot TRUE/FALSE}

\item{intrans}{calculate intransitivities (calculation intense!)}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{ordn}{item category order}

\item{r1}{label of the test item (e.g., "Lake")}

\item{r2}{label(s) of the remaining item(s)}

\item{ymin}{minimum y-scale of the worth plot}

\item{ymax}{maximum y-scale of the worth plot}
}
\value{
Exports the results of the worth value calculation, including the GNM analysis.
}
\description{
The \code{tie_import} function prepares binary and continuous data form import into the tiefightR analysis.
}
