% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_binarize.R
\name{tie_binarize}
\alias{tie_binarize}
\title{Binarize continous data}
\usage{
tie_binarize(
  xdata = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  datalabel = "binarized",
  compiled_studies = NULL,
  setseed = TRUE,
  prefLimit = 50,
  refval = "refValue",
  oval = "otherValue",
  aid = "animalID",
  fldrnk = "fluid_drunk"
)
}
\arguments{
\item{xdata}{imported (binarized) data frame}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{datalabel}{universal study label for the binarized data}

\item{compiled_studies}{label of the compiled sub study (used for filtering)}

\item{setseed}{TRUE/FALSE for seeding}

\item{prefLimit}{preference limit for binarization threshold}

\item{refval}{the name of the standardized reference variable in the org data}

\item{oval}{the name of the standardized combination variable in the org data}

\item{aid}{the name of the standardized animal id column in the org data}

\item{fldrnk}{the name of the standardizedresponse variable in the org data}
}
\value{
binarized data in the default format (colum headers)
}
\description{
Uses \code{tie_import} non-binary outcome data as input. The outcome variable is randomized for ties in the continous data.
}
