% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie_test.R
\name{tie_test}
\alias{tie_test}
\title{Test Function}
\usage{
tie_test(
  xdata = NULL,
  R = 2,
  SV = NULL,
  RF = NULL,
  CF = NULL,
  id = NULL,
  RV = NULL,
  intrans = TRUE,
  compstudy = NULL,
  default = NULL,
  ord = NULL,
  seed = TRUE,
  testme = NULL,
  against = NULL
)
}
\arguments{
\item{xdata}{imported (binarized) data frame}

\item{SV}{name of the side variable}

\item{RF}{name of the reference fluid variable}

\item{CF}{name of the combination fluid variable}

\item{id}{subject IDs}

\item{RV}{name of the response variable}

\item{intrans}{calculate intransitivities (calculation intense!)}

\item{compstudy}{label of the compiled sub study (used for filtering)}

\item{default}{default item in worth value estimation (usually the lowest worth value)}

\item{esti}{worth estimator (default, "worth", alt: "estimator")}

\item{r1}{label of the test item (e.g., "Lake")}

\item{r2}{label(s) of the remaining item(s)}

\item{showplot}{show worth plot TRUE/FALSE}

\item{ymax}{maximum y-scale of the worth plot}

\item{ordn}{item category order}
}
\value{
A frequency table for item positions during the simulation; Tukey's HSD Test for positions; Position Bubble Plot
}
\description{
The \code{tie_test} function can be used for individual item testing.
}
